Multiple features: - Improve history loading, formatting and logging - Add prompt when clearing history

+++ tfmttools-cli/src/commands/clear_history.rs
@@ -3,24 +3,39 @@ use fs_err as fs;
 
 use super::Command;
 use crate::config::{Config, DRY_RUN_PREFIX};
+use crate::history::{load_history, HistoryFormatter};
+use crate::ui::ConfirmationPrompt;
 
 #[derive(Debug)]
 pub struct ClearHistory;
-
-// TODO? Prompt for confirmation?
 
 impl Command for ClearHistory {
     fn run(&self, config: &Config) -> Result<()> {
         let path = &config.history_file();
 
         if path.is_file() {
-            if config.dry_run() {
-                print!("{DRY_RUN_PREFIX}");
-            } else {
-                fs::remove_file(path)?;
-            }
+            let result = load_history(config)?;
 
-            println!("Removed history file at: {path}");
+            let history = result.unwrap();
+
+            let formatter = HistoryFormatter::normal();
+
+            println!("{}", formatter.format(&history));
+
+            let confirmation =
+                ConfirmationPrompt::new("Clear history?").prompt()?;
+
+            if confirmation {
+                if config.dry_run() {
+                    print!("{DRY_RUN_PREFIX}");
+                } else {
+                    fs::remove_file(path)?;
+                }
+
+                println!("Removed history file at: {path}");
+            } else {
+                println!("Aborting.");
+            }
         } else if path.exists() {
             eprintln!("History file path exists, but is not a file: {path}");
         } else {
